save ifw rules to specified folder
